+++ src/test/scala/mesosphere/marathon/test/MarathonTestHelper.scala
@@ -50,7 +50,7 @@ object MarathonTestHelper {
   }
 
   def defaultConfig(
-    maxTasksPerOffer: Int = 1,
+    maxInstancesPerOffer: Int = 1,
     minReviveOffersInterval: Long = 100,
     mesosRole: Option[String] = None,
     acceptedResourceRoles: Option[Set[String]] = None,
@@ -60,7 +60,7 @@ object MarathonTestHelper {
 
     var args = Seq(
       "--master", "127.0.0.1:5050",
-      "--max_tasks_per_offer", maxTasksPerOffer.toString,
+      "--max_instances_per_offer", maxInstancesPerOffer.toString,
       "--min_revive_offers_interval", minReviveOffersInterval.toString,
       "--mesos_authentication_principal", "marathon"
     )